jan 20

+++ frontend/src/utils/Api.js
@@ -42,16 +42,19 @@ class Api {
     }
 
     changeLikeCardStatus (cardID, like) {
-        return ((like ? fetch(this._baseUrl + '/cards/' + cardID + '/likes', {
-            headers: this._headers,
-            method: "PUT"
-        })
-            : fetch(this._baseUrl + '/cards/' + cardID + '/likes', {
-            headers: this._headers,
-            method: "DELETE"
-        }))
+        if (like) {
+            return fetch(this._baseUrl + '/cards/' + cardID + '/likes', {
+                headers: this._headers,
+                method: "DELETE"
+            })
             .then(res => res.ok ? res.json() : Promise.reject('Error' + res.statusText))
-        );
+        } else {
+            return fetch(this._baseUrl + '/cards/' + cardID + '/likes', {
+                headers: this._headers,
+                method: "PUT"
+            })
+            .then(res => res.ok ? res.json() : Promise.reject('Error' + res.statusText))
+        }
     }
 
     setUserInfo(data) {
